feat: refactor the file directory
