+++ js-plugins/postcss/farm.config.mjs
@@ -4,8 +4,13 @@ import { createFarmJsPluginBuildConfig } from '../../configs/farm-js-plugin.base
 /**
  * @type {import('@farmfe/core').UserConfig}
  */
-export default createFarmJsPluginBuildConfig([
-  farmDtsPlugin({
-    tsConfigPath: './tsconfig.build.json'
-  })
-]);
+export default createFarmJsPluginBuildConfig(
+  [
+    farmDtsPlugin({
+      tsConfigPath: './tsconfig.build.json'
+    })
+  ],
+  {
+    external: ['@farmfe/core']
+  }
+);
