+++ packages/core/src/utils/file.ts
@@ -1,3 +1,8 @@
+import fs from 'fs';
+import fsp from 'node:fs/promises';
+import { normalizePath } from './share.js';
+import path from 'path';
+
 interface FileNode {
   isLeaf: boolean;
   name: string;
@@ -67,4 +72,36 @@ export function generateFileTreeHtml(node: FileNode[]): string {
         </body>
       </html>
       `;
+}
+
+export const ERR_SYMLINK_IN_RECURSIVE_READDIR =
+  'ERR_SYMLINK_IN_RECURSIVE_READDIR';
+export async function recursiveReaddir(dir: string): Promise<string[]> {
+  if (!fs.existsSync(dir)) {
+    return [];
+  }
+  let dirents: fs.Dirent[];
+  try {
+    dirents = await fsp.readdir(dir, { withFileTypes: true });
+  } catch (e) {
+    if (e.code === 'EACCES') {
+      // Ignore permission errors
+      return [];
+    }
+    throw e;
+  }
+  if (dirents.some((dirent) => dirent.isSymbolicLink())) {
+    const err: any = new Error(
+      'Symbolic links are not supported in recursiveReaddir'
+    );
+    err.code = ERR_SYMLINK_IN_RECURSIVE_READDIR;
+    throw err;
+  }
+  const files = await Promise.all(
+    dirents.map((dirent) => {
+      const res = path.resolve(dir, dirent.name);
+      return dirent.isDirectory() ? recursiveReaddir(res) : normalizePath(res);
+    })
+  );
+  return files.flat(1);
 }
