+++ packages/core/src/utils/index.ts
@@ -3,3 +3,6 @@ export * from './logger.js';
 export * from './build.js';
 export * from './file.js';
 export * from './color.js';
+export * from './url.js';
+export * from './path.js';
+export * from './publicDir.js';
