+++ packages/core/src/utils/path.ts
@@ -1,0 +1,11 @@
+const windowsSlashRE = /\\/g;
+export function slash(p: string): string {
+  return p.replace(windowsSlashRE, '/');
+}
+
+export function withTrailingSlash(path: string): string {
+  if (path[path.length - 1] !== '/') {
+    return `${path}/`;
+  }
+  return path;
+}
