+++ packages/core/src/utils/url.ts
@@ -1,0 +1,4 @@
+const postfixRE = /[?#].*$/;
+export function cleanUrl(url: string): string {
+  return url.replace(postfixRE, '');
+}
